Fixes #8: Removed duplicate classes and consolidated admin/user login functionality
