--- src/main/java/com/example/demo/loginCredentials/AdminLogin.java
@@ -1,26 +1,0 @@
-package com.example.demo.loginCredentials;
-
-public class AdminLogin
-{
-	
-	private String email;
-	
-	private String password;
-	
-	public String getEmail() {
-		return email;
-	}
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	public String getPassword() {
-		return password;
-	}
-	public void setPassword(String password) {
-		this.password = password;
-	}
-	@Override
-	public String toString() {
-		return "AdminLogin [name=" + email + ", password=" + password + "]";
-	}	
-}
